requirements.txt
python-telegram-bot==21.9
openai==1.68.0
python-dotenv==1.0.1
